Fix random seed

+++ arduino_protocols/platform_avoidance/PLAT_AVOID_W_LP_LOG/PLAT_AVOID_W_LP_LOG.ino
@@ -170,7 +170,7 @@ void setup() {
   pinMode(6, OUTPUT);  // DElIVER US
 
   // ENSURE REPRODUCIBILITY
-  unsigned long seed = 31;
+  int seed = 31;
   randomSeed(seed);
 
 }
